Create README.md
A computer vision project.
I'll be learning how to change a computer's volume using Gesture Control in this project. I first research hand tracking, and then I use hand landmarks to identify a hand gesture that changes the volume. As this project is module-based, I will be using a hand module , which makes the hand tracking really simple.